+++ packages/eas-cli/src/commands/branch/view.ts
@@ -1,17 +1,10 @@
-import { Flags } from '@oclif/core';
-import chalk from 'chalk';
-import Table from 'cli-table3';
-
+import { selectBranchFromPaginatedQueryAsync } from '../../branch/queries';
 import EasCommand from '../../commandUtils/EasCommand';
-import { UpdateQuery } from '../../graphql/queries/UpdateQuery';
-import Log from '../../log';
+import { EasPaginatedQueryFlags, getPaginatedQueryOptions } from '../../commandUtils/pagination';
 import { getExpoConfig } from '../../project/expoConfig';
 import { findProjectRootAsync, getProjectIdAsync } from '../../project/projectUtils';
-import { promptAsync } from '../../prompts';
-import { UPDATE_COLUMNS, formatUpdate, getPlatformsForGroup } from '../../update/utils';
-import groupBy from '../../utils/expodash/groupBy';
-import formatFields from '../../utils/formatFields';
-import { enableJsonOutput, printJsonOnlyOutput } from '../../utils/json';
+import { listAndRenderUpdatesOnBranchByNameAsync } from '../../update/queries';
+import { enableJsonOutput } from '../../utils/json';
 
 export default class BranchView extends EasCommand {
   static description = 'view a branch';
@@ -25,18 +18,17 @@ export default class BranchView extends EasCommand {
   ];
 
   static flags = {
-    json: Flags.boolean({
-      description: `return a json with the branch's ID name and recent update groups.`,
-      default: false,
-    }),
+    ...EasPaginatedQueryFlags,
   };
 
   async runAsync(): Promise<void> {
     let {
-      args: { name },
-      flags: { json: jsonFlag },
+      args: { name: branchName },
+      flags,
     } = await this.parse(BranchView);
-    if (jsonFlag) {
+    const options = getPaginatedQueryOptions(flags);
+
+    if (options.json) {
       enableJsonOutput();
     }
 
@@ -44,63 +36,20 @@ export default class BranchView extends EasCommand {
     const exp = getExpoConfig(projectDir);
     const projectId = await getProjectIdAsync(exp);
 
-    if (!name) {
-      const validationMessage = 'Branch name may not be empty.';
-      if (jsonFlag) {
-        throw new Error(validationMessage);
-      }
-      ({ name } = await promptAsync({
-        type: 'text',
-        name: 'name',
-        message: 'Please enter the name of the branch to view:',
-        validate: value => (value ? true : validationMessage),
-      }));
-    }
-
-    const { app } = await UpdateQuery.viewBranchAsync({
-      appId: projectId,
-      name,
-    });
-    const UpdateBranch = app?.byId.updateBranchByName;
-    if (!UpdateBranch) {
-      throw new Error(`Could not find branch "${name}"`);
-    }
-
-    const updates = Object.values(groupBy(UpdateBranch.updates, u => u.group)).map(
-      group => group[0]
-    );
-
-    if (jsonFlag) {
-      printJsonOnlyOutput({ ...UpdateBranch, updates });
-    } else {
-      const groupTable = new Table({
-        head: UPDATE_COLUMNS,
-        wordWrap: true,
-      });
-
-      for (const update of updates) {
-        groupTable.push([
-          formatUpdate(update),
-          update.runtimeVersion,
-          update.group,
-          getPlatformsForGroup({
-            updates: UpdateBranch.updates,
-            group: update.group,
-          }),
-        ]);
+    // provide help to a user if they ran the command with missing args
+    if (!branchName) {
+      if (options.nonInteractive) {
+        throw new Error('Branch name may not be empty.');
       }
 
-      Log.addNewLineIfNone();
-      Log.log(chalk.bold('Branch:'));
-      Log.log(
-        formatFields([
-          { label: 'Name', value: UpdateBranch.name },
-          { label: 'ID', value: UpdateBranch.id },
-        ])
-      );
-      Log.addNewLineIfNone();
-      Log.log(chalk.bold('Recently published update groups:'));
-      Log.log(groupTable.toString());
+      ({ name: branchName } = await selectBranchFromPaginatedQueryAsync(
+        projectId,
+        'Which branch would you like to view?',
+        // discard limit and offset because this query is not those flag's intended target
+        { json: options.json, nonInteractive: options.nonInteractive, offset: 0 }
+      ));
     }
+
+    await listAndRenderUpdatesOnBranchByNameAsync(projectId, branchName, options);
   }
 }